requirements.txt
fastapi~=0.121.2

SQLAlchemy~=2.0.44
alembic~=1.14.1
pydantic~=2.12.4
passlib~=1.7.4
python-dotenv~=1.2.1
python-jose~=3.5.0
bcrypt~=4.3.0
pydantic[email]
mysql
mysql-connector-python
python-multipart
pymongo~=4.15.4

httpx~=0.28.1
pydantic_core~=2.41.5
starlette~=0.49.3